# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vacation Optimizer",
    "version": "0.1.0"
  },
  "paths": {
...
    "/holidays/{year}": {
      "get": {
        "summary": "Holidays",
        "description": "Observed federal holidays that fall in `year` (including next year's\nNew Year's when it's observed on Dec 31).",
        "operationId": "holidays_holidays__year__get",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Holiday"
                  },
                  "title": "Response Holidays Holidays  Year  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me": {
      "get": {
        "summary": "Me",
        "description": "Who's signed in, the people they manage, and the groups they can open.",
        "operationId": "me_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/people": {
      "post": {
        "summary": "Create Person",
        "operationId": "create_person_people_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/people/{person_id}/calendars": {
      "post": {
        "summary": "Subscribe",
        "description": "Turn a calendar on for a person (or update their excluded holidays).",
        "operationId": "subscribe_people__person_id__calendars_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Person Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeIn"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/people/{person_id}/pto-blocks": {
      "post": {
        "summary": "Add Pto Block",
        "operationId": "add_pto_block_people__person_id__pto_blocks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Person Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PTOBlockIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PTOBlockOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/calendars": {
      "get": {
        "summary": "List Calendars",
        "description": "Shared calendars plus this account's own.",
        "operationId": "list_calendars_calendars_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CalendarOut"
                  },
                  "type": "array",
                  "title": "Response List Calendars Calendars Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "summary": "Create Calendar",
        "operationId": "create_calendar_calendars_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/calendars/{calendar_id}/events": {
      "post": {
        "summary": "Add Event",
        "operationId": "add_event_calendars__calendar_id__events_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "calendar_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Calendar Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/groups": {
      "post": {
        "summary": "Create Group",
        "operationId": "create_group_groups_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/groups/{group_id}": {
      "get": {
        "summary": "Get Group",
        "operationId": "get_group_groups__group_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Group Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/groups/{group_id}/members/{person_id}": {
      "delete": {
        "summary": "Remove Member",
        "description": "The owner can remove anyone; a member can take their own people out.",
        "operationId": "remove_member_groups__group_id__members__person_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Group Id"
            }
          },
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Person Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/groups/{group_id}/windows": {
      "get": {
        "summary": "Optimize",
        "description": "Rank the vacation windows everyone in the group can share.",
        "operationId": "optimize_groups__group_id__windows_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Group Id"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "End"
            }
          },
          {
            "name": "min_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 31,
              "minimum": 2,
              "default": 3,
              "title": "Min Days"
            }
          },
          {
            "name": "max_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 31,
              "minimum": 2,
              "default": 16,
              "title": "Max Days"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "best_value",
                "longest"
              ],
              "type": "string",
              "default": "best_value",
              "title": "Sort"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "distinct",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Distinct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CalendarIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "kind": {
            "$ref": "#/components/schemas/CalendarKind"
          }
        },
        "type": "object",
        "required": [
          "name",
          "kind"
        ],
        "title": "CalendarIn"
      },
...
      "CalendarOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "$ref": "#/components/schemas/CalendarKind"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "kind"
        ],
        "title": "CalendarOut"
      },
...
      "EventIn": {
        "properties": {
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "effect": {
            "$ref": "#/components/schemas/Effect"
          }
        },
        "type": "object",
        "required": [
          "start_date",
          "end_date",
          "title",
          "effect"
        ],
        "title": "EventIn"
      },
      "EventOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "effect": {
            "$ref": "#/components/schemas/Effect"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "start_date",
          "end_date",
          "effect"
        ],
        "title": "EventOut"
      },
      "GroupIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "person_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "minItems": 1,
            "title": "Person Ids"
          }
        },
        "type": "object",
        "required": [
          "name",
          "person_ids"
        ],
        "title": "GroupIn"
      },
      "GroupOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "MEMBER"
            ],
            "title": "Role"
          },
          "people": {
            "items": {
              "$ref": "#/components/schemas/MemberOut"
            },
            "type": "array",
            "title": "People"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "role",
          "people"
        ],
        "title": "GroupOut"
      },
      "GroupSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "MEMBER"
            ],
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "role"
        ],
        "title": "GroupSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Holiday": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "date",
          "name"
        ],
        "title": "Holiday"
      },
...
      "MeOut": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "self_person_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Self Person Id"
          },
          "people": {
            "items": {
              "$ref": "#/components/schemas/PersonOut"
            },
            "type": "array",
            "title": "People"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/GroupSummary"
            },
            "type": "array",
            "title": "Groups"
          }
        },
        "type": "object",
        "required": [
          "email",
          "name",
          "self_person_id",
          "people",
          "groups"
        ],
        "title": "MeOut"
      },
...
      "PTOBlockIn": {
        "properties": {
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "status": {
            "$ref": "#/components/schemas/PTOStatus"
          }
        },
        "type": "object",
        "required": [
          "start_date",
          "end_date",
          "status"
        ],
        "title": "PTOBlockIn"
      },
      "PTOBlockOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "status": {
            "$ref": "#/components/schemas/PTOStatus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "start_date",
          "end_date",
          "status"
        ],
        "title": "PTOBlockOut"
      },
      "PTOStatus": {
        "type": "string",
        "enum": [
          "PROPOSED",
          "COMMITTED"
        ],
        "title": "PTOStatus"
      },
      "PersonIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "kind": {
            "$ref": "#/components/schemas/PersonKind"
          },
          "pto_balance": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Pto Balance"
          },
          "work_week": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Work Week",
            "description": "Mon=0 ... Sun=6",
            "default": [
              0,
              1,
              2,
              3,
              4
            ]
          },
          "is_self": {
            "type": "boolean",
            "title": "Is Self",
            "description": "This person is the signed-in user",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "kind"
        ],
        "title": "PersonIn"
      },
      "PersonKind": {
        "type": "string",
        "enum": [
          "ADULT",
          "CHILD",
          "GUEST"
        ],
        "title": "PersonKind"
      },
      "PersonOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "$ref": "#/components/schemas/PersonKind"
          },
          "pto_balance": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pto Balance"
          },
          "work_week": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Work Week"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "kind",
          "pto_balance",
          "work_week"
        ],
        "title": "PersonOut"
      },
      "SearchOut": {
        "properties": {
          "windows": {
            "items": {
              "$ref": "#/components/schemas/WindowOut"
            },
            "type": "array",
            "title": "Windows"
          },
          "free_long_weekends": {
            "items": {
              "$ref": "#/components/schemas/WindowOut"
            },
            "type": "array",
            "title": "Free Long Weekends"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "windows",
          "free_long_weekends",
          "message"
        ],
        "title": "SearchOut"
      },
...
      "SubscribeIn": {
        "properties": {
          "calendar_id": {
            "type": "integer",
            "title": "Calendar Id"
          },
          "excluded_titles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Excluded Titles",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "calendar_id"
        ],
        "title": "SubscribeIn"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}